Refactoring project package
- Add proguard rules example
- Add ViewPagerFixed (Issue in : https://github.com/chrisbanes/PhotoView/issues/31#issuecomment-19803926)
