+++ app/src/main/java/kr/co/bluebright/www/myexperiment/common/util/PreferenceController.java
@@ -1,4 +1,4 @@
-package kr.co.bluebright.www.myexperiment.pref;
+package kr.co.bluebright.www.myexperiment.common.util;
 
 import android.content.Context;
 import android.content.SharedPreferences;
