+++ app/src/main/java/kr/co/bluebright/www/myexperiment/core/MyApplication.java
@@ -1,4 +1,4 @@
-package kr.co.bluebright.www.myexperiment.app;
+package kr.co.bluebright.www.myexperiment.core;
 
 import android.app.Application;
