+++ app/src/main/res/layout/activity_splash.xml
@@ -6,7 +6,7 @@
     <androidx.constraintlayout.widget.ConstraintLayout
         android:layout_width="match_parent"
         android:layout_height="match_parent"
-        tools:context=".SplashActivity">
+        tools:context=".intro.SplashActivity">
 
         <com.facebook.drawee.view.SimpleDraweeView
             android:id="@+id/img_logo"
