+++ app/src/main/res/layout/content_main.xml
@@ -6,7 +6,7 @@
     <androidx.constraintlayout.widget.ConstraintLayout
         android:layout_width="match_parent"
         android:layout_height="match_parent"
-        tools:context=".MainActivity"
+        tools:context=".main.MainActivity"
         tools:showIn="@layout/activity_main">
 
         <TextView
