+++ proguard-rules-admob.pro
@@ -1,0 +1,11 @@
+##---------------Begin: proguard configuration for AdMob  ----------
+
+-keep public class com.gogole.android.gms.ads.**{
+    public *;
+}
+
+-keep public class com.google.ads.** {
+    public *;
+}
+
+##---------------End: proguard configuration for AdMob  ----------
