+++ proguard-rules-android-youtube-player.pro
@@ -1,0 +1,22 @@
+# GoogleApiClient
+# Needed to keep generic types and @Key annotations accessed via reflection
+-keepattributes Signature,RuntimeVisibleAnnotations,AnnotationDefault
+-keepclassmembers class * {
+  @com.google.api.client.util.Key <fields>;
+}
+# Needed by google-http-client-android when linking against an older platform version
+-dontwarn com.google.api.client.extensions.android.**
+# Needed by google-api-client-android when linking against an older platform version
+-dontwarn com.google.api.client.googleapis.extensions.android.**
+# Needed by google-play-services when linking against an older platform version
+-dontwarn com.google.android.gms.**
+# com.google.client.util.IOUtils references java.nio.file.Files when on Java 7+
+-dontnote java.nio.file.Files, java.nio.file.Path
+# Suppress notes on LicensingServices
+-dontnote **.ILicensingService
+# Suppress warnings on sun.misc.Unsafe
+-dontnote sun.misc.Unsafe
+-dontwarn sun.misc.Unsafe
+
+-dontwarn com.google.common.**
+-dontwarn com.google.api.client.json.**
