+++ proguard-rules-firebase.pro
@@ -1,0 +1,9 @@
+# Begin: Proguard rules for Firebase
+
+# Authentication
+-keepattributes *Annotation*
+
+# Realtime database
+-keepattributes Signature
+
+# End: Proguard rules for Firebase
