+++ proguard-rules-gson.pro
@@ -1,0 +1,37 @@
+
+# Add any classes the interact with gson
+-keepclassmembers class com.madx.quiz.apf.apf.APFQuestion { *; }
+-keepclassmembers class com.madx.quiz.apf.apf.APFSubCategory { *; }
+
+##---------------Begin: proguard configuration for Gson  ----------
+
+# Gson uses generic type information stored in a class file when working with fields. Proguard
+# removes such information by default, so configure it to keep all of it.
+-keepattributes Signature
+
+# For using GSON @Expose annotation
+-keepattributes *Annotation*
+
+# Gson specific classes
+-dontwarn sun.misc.**
+#-keep class com.google.gson.stream.** { *; }
+
+# Application classes that will be serialized/deserialized over Gson
+-keep class com.google.gson.examples.android.model.** { <fields>; }
+
+# Prevent proguard from stripping interface information from TypeAdapterFactory,
+# JsonSerializer, JsonDeserializer instances (so they can be used in @JsonAdapter)
+-keep class * implements com.google.gson.TypeAdapterFactory
+-keep class * implements com.google.gson.JsonSerializer
+-keep class * implements com.google.gson.JsonDeserializer
+
+# Prevent R8 from leaving Data object members always null
+-keepclassmembers,allowobfuscation class * {
+  @com.google.gson.annotations.SerializedName <fields>;
+}
+
+##---------------End: proguard configuration for Gson  ----------
+
+# Ignore Gson warning
+-dontnote sun.misc.Unsafe
+-dontnote com.google.gson.internal.UnsafeAllocator
