+++ proguard-rules-okhttp3.pro
@@ -1,0 +1,8 @@
+# Begin: Proguard rules for okhttp3
+
+-dontwarn okhttp3.**
+-dontwarn okio.**
+
+-dontnote okhttp3.**
+
+# End: Proguard rules for okhttp3
