+++ proguard-rules-picasso.pro
@@ -1,0 +1,3 @@
+##---------------Begin: proguard configuration for Picasso  ----------
+-dontwarn com.squareup.okhttp.**
+##---------------End: proguard configuration for Picasso  ----------
